Avoid window.open for coder apps
The current implementation of `AppLink` opens the specified application
in a small popup instead of a new browser tab. The experience is not
good because you lose all the browser toolbars. Also, if you open
external links like `vscode://`, VSCode is opened but an empty browser
window is left as leftover.