Use modifier properties on keydown event
Instead of manually tracking modifiers with keydown/keyup, as this can
miss keys like in the case of ctrl+tab, leaving ctrl in a permanent "on"
state.

The other modifiers do not appear to be used in conjunction with other
keys, except possibly Fn?  So this should be more or less the same
functionality.