feat(cli): integrate chromedp actions into scaletest dashboard
This commit integrates the chromedp actions in the previous commit
into the scaletest dashboard command, and re-enables the previously
disabled unit test.

Note that this unit test does not actually run headless chrome, nor
does it test the actual scaletest actions yet, as coderdtest only
exposes an API and does not expose the actual site.