fix: find jetbrains on ipv6 when port collides
This commit was mostly just to add a test for ipv6, but in so doing I
discovered it was possible to miss a JetBrains process listening on ipv6
if there was something else non-JetBrains on ipv4 with the same port.
This is fixed now, so we should never miss processes (aside from
errors).

On the flip side, while it was already possible to get false positives
if JetBrains was on ipv4 and something else was forwarded on ipv6, it is
now possible to get them the other way around too.  I am not sure it is
worth trying to resolve these false positives as producing them requires
a setup that seems quite unlikely and the impact is low.