chore(dogfood): fix datadog trace exporter
The DataDog tracer provider was only doing runtime metrics and wasn't
attached to the global otel tracer provider. This changes tracing so if
DataDog is enabled, it will become the sole tracer provider.

If other providers are set (CODER_TRACE_ENABLE=true or
CODER_TRACE_HONEYCOMB_API_KEY=xxx), we return an error on start because
they conflict.

This avoids us having to use an OTEL endpoint in dogfood and lets us
just use DataDog for everything.