fix(coderd): remove websocket wait group on shutdown
There's really no great way to shutdown websockets. In the code
currently it seems to assume that `(*http.Server).Shutdown()`
gracefully kills them, but that's not the case: https://pkg.go.dev/net/http#Server.Shutdown

This removes the mutex so that `coder server` no longer waits for them
to exit.