feat(cli): add auth command (#11004)
Adds `coder auth` command with 3 sub-commands: `coder auth token`,
`coder auth status` and `coder auth login`.

`coder auth login`: This is the same command as `coder login`.

`coder auth status`: Informs the user about their authentication
status. It may make sense for this to not show an error
if called when user is not logged in.

`coder auth token`: Informs the user of their token string
and the expiration time of the token.

All the commands use a middleware to check that the user is
authenticated and will return an error if the user is not. They also
make at most 2 API calls.

The commands have additional values that could be shown but are not with
this implementation.