Show ineditable form without org edit perm
I think the "you do not have permission to edit" message might make it
sound like you can do nothing at all with the org, but it might be that
you can edit members and groups, just not the org settings.

So, instead, show the org settings but disable the form if you cannot
edit.  Probably there is still a better design but maybe this is OK for
now.