Limit the number of org permission checks
We show the org on the sidebar if they can edit anything, and we show
each sub-link if they can view it, which means we were making both edit
and view permission checks.

Instead, show each link if they can edit it (not just view), which
negates the need for separate view permissions.

Incidentally, this also reduces the number of checks we need to make for
individual pages, since some of them were only used on the sidebar.