feat: add --network-info-dir and --network-info-interval flags to coder ssh
This is the first in a series of PRs to enable "coder ssh" to replace
"coder vscodessh".

This change adds --network-info-dir and --network-info-interval flags to
the ssh subcommand. These were formerly only available with the
vscodessh subcommand.

Subsequent PRs will add a --ssh-host-prefix flag to the ssh subcommand,
and adjust the log file naming to contain the parent PID.