feat: parse structured hostname arguments to "coder ssh"
If the argument to `coder ssh` fits the pattern
`<prefix>--<owner>--<workspace>--<agent?>`, then parse this instead of
treating the argument as a literal workspace name or
`<owner>/<workspace>`.

This will enable the use of `coder ssh` for the VS Code and JetBrains
plugins in combination with wildcard `Host` entries in the SSH config.

Part of #14986.