chore(coderd/rbac): add `Action{Create,Delete}Agent` to `ResourceWorkspace`
This PR adds two new actions: `ActionCreateAgent` and
`ActionDeleteAgent`. The former is used in this PR for
`InsertWorkspaceAgent`, with the latter to be used in a follow-up PR for
Dev Container Agents.

A note has been left in `dbauthz.go` for `InsertWorkspaceAgent`
detailing why it is allowed to insert a workspace agent when no
workspace could be found.