fix: for child agents, ensure unique across provisioner job
Child agents are a new concept so there are very little existing tests
for them. This means we can easily make the change for them without
having to update a mountain of tests. So they get a special case for
checking properly, whilst parent agents have less scrutiny due to the
vast number of tests that appear to not work properly.