chore(coderd/database/dbauthz): update RBAC for InsertWorkspaceApp
Instead of using `ResourceSystem` as the resource for
`InsertWorkspaceApp`, we instead use the associated workspace (if it
exists), with the action `ActionCreateAgent`. The decision to use this
action is because everywhere where `InsertWorkspaceApp` is called is
also where `InsertWorkspaceAgent` is called. I don't think adding an
extra action makes sense, and adding an extra resource also doesn't make
sense.