Use common flags for agent client in MCP server
Before we were separately reading the environment, but we can reuse the
global flags and the function for creating an agent client.

Also, since the tests no longer need to set environment variables they
can now be ran in parallel.

This is not fully backwards compatible in that the old method would fall
back to the non-agent client URL if the agent URL was not set, but that
seems like undesired behavior since we are not doing that anywhere else.