Move update predicates to push phase
Instead of the pop phase.  This ensures we do not queue up updates that
will just end up being discarded once they are popped (which could take
some time due to latency to coderd).

It also has the side effect of preserving summaries even when the queue
gets too big, because now we preserve them as part of pushing, before
they might get lost due to getting dropped while we wait on coderd.