fix(cli)!: select sub agent if available and error on multiple agents
In the past we randomly selected workspace agent if there were multiple.
Unless both are running on the same machine with the same configuration,
this would be very confusing behavior for a user.

With the introduction of sub agents (devcontainer agents), it was
decided prioritize them if present. Similarly as before, selecting a
devcontainer randomly would be confusing. We now error out if the agent
name is not specified and there are multiple agents.

Fixes coder/internal#696