refactor: include configuration and merged configuration
We want to be able to pick customizations that are only specified in the
top-level customizations stanza. That is, we do not want to accidentally
pick something specified in a feature (such as the agent name).

This refactor picks the agent name from configuration instead of the
merged configuration to ensure a feature cannot influence it.