chore(agent/agentcontainers): skip part of test if on `darwin`
Relates to https://github.com/coder/internal/issues/765

I've been unable to figure out how to fix this (unfortunately my macOS
file system knowledge is lacking). Given that this test is causing a lot
of flakes at the moment, I think it is best to skip this part of the
test on macOS for now.